+++ lib/THNN/generic/SpatialFullConvolutionMap.c
@@ -4,8 +4,8 @@
 
 void THNN_(SpatialFullConvolutionMap_updateOutput)(
   THNNState *state, THTensor *input, THTensor *output_, THTensor *weight, THTensor *bias,
-  THTensor *connTable, int nInputPlane, int nOutputPlane,
-  int dW, int dH)
+  THTensor *connTable, long nInputPlane, long nOutputPlane,
+  long dW, long dH)
 {
   THArgCheck(
     weight != NULL && weight->nDimension == 3
@@ -85,8 +85,8 @@ void THNN_(SpatialFullConvolutionMap_updateOutput)(
 
 void THNN_(SpatialFullConvolutionMap_updateGradInput)(
   THNNState *state, THTensor *input, THTensor *gradOutput, THTensor *gradInput_, THTensor *weight, THTensor *bias,
-  THTensor *connTable, int nInputPlane, int nOutputPlane,
-  int dW, int dH)
+  THTensor *connTable, long nInputPlane, long nOutputPlane,
+  long dW, long dH)
 {
   THArgCheck(
     weight != NULL && weight->nDimension == 3
@@ -153,9 +153,9 @@ void THNN_(SpatialFullConvolutionMap_accGradParameters)(
   THTensor *gradWeight,
   THTensor *gradBias,
   THTensor *connTable,
-  int nInputPlane,
-  int nOutputPlane,
-  int dW, int dH,
+  long nInputPlane,
+  long nOutputPlane,
+  long dW, long dH,
   accreal scale_)
 {
   real scale = TH_CONVERT_ACCREAL_TO_REAL(scale_);
